refactor: rest_api_client to pub_dev_api_client
