+++ packages/flutter_app/lib/features/pub_dev_packages/query/pub_dev_packages_page_state_provider.dart
@@ -1,7 +1,7 @@
-import 'package:rest_api_client/rest_api_client.dart';
+import 'package:pub_dev_api_client/pub_dev_api_client.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../../package_adaptor/rest_api_client_provider.dart';
+import '../../../package_adaptor/pub_dev_api_client_provider.dart';
 
 part 'pub_dev_packages_page_state_provider.g.dart';
