+++ packages/flutter_app/lib/package_adaptor/pub_dev_api_client_provider.dart
@@ -1,12 +1,12 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:rest_api_client/rest_api_client.dart';
+import 'package:pub_dev_api_client/pub_dev_api_client.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'rest_api_client_provider.g.dart';
 
 @Riverpod(keepAlive: true)
-RestApiClient restApiClient(Ref ref) {
-  return RestApiClient.fromParameters(
+PubDevApiClient pubDevApiClient(Ref ref) {
+  return PubDevApiClient.fromParameters(
     baseUrl: 'https://pub.dev/api',
     timeoutDuration: const Duration(seconds: 30),
   );
