+++ packages/flutter_app/pubspec.yaml
@@ -43,8 +43,8 @@ dependencies:
   json_annotation: ^4.8.1
   linked_scroll_controller: ^0.2.0
   package_info_plus: ^8.0.2
-  rest_api_client:
-    path: ../rest_api_client
+  pub_dev_api_client:
+    path: ../pub_dev_api_client
   riverpod_annotation: ^2.6.1
   shared_preferences: ^2.3.2
   simple_logger: ^1.9.0+1
