+++ packages/pub_dev_api_client/lib/pub_dev_api_client.dart
@@ -1,0 +1,5 @@
+/// Library that provides a client for the Pub.dev API.
+library;
+
+export 'src/models/models.dart';
+export 'src/pub_dev_api_client.dart';
