+++ packages/pub_dev_api_client/lib/src/models/get_package_details_response_body.dart
