+++ packages/pub_dev_api_client/lib/src/models/latest_package_release.freezed.dart
