+++ packages/pub_dev_api_client/lib/src/models/package_name.freezed.dart
