--- packages/rest_api_client/lib/rest_api_client.dart
@@ -1,5 +1,0 @@
-/// Library that provides a client for the REST API.
-library;
-
-export 'src/models/models.dart';
-export 'src/rest_api_client.dart';
